add CHANGELOG.md, refactor code, and update action.yml
